Add support for i2c and platform binding
Some Platfomrs are not listing i2c as resource in ACPI tables
this is expected due to some CVS FW not supporting i2c communication.

The driver should support both and perform only GPIO based operations if
i2c is not present, if i2c is present, the full features are enabled.

Signed-off-by: Miguel Vadillo <miguel.vadillo@intel.com>